fixed inconsistency when calling the Http Kernel instance from an event
